refactor: move http to graphql types module
