+++ src/blueprint/from_config/graphql/mod.rs
@@ -1,5 +1,7 @@
 mod schema;
 mod union;
+mod types;
 
 pub use schema::*;
 pub use union::*;
+pub use types::*;
