+++ src/blueprint/from_config/graphql/types/mod.rs
@@ -1,0 +1,3 @@
+mod http;
+
+pub use http::*;
